fix: 상세 페이지 공통 훅을 개별 페이지 훅으로 변경
