+++ src/pages/projects/[projectId].tsx
@@ -2,10 +2,10 @@ import styled from '@emotion/styled';
 import { Box, Divider, Stack, Typography } from '@mui/material';
 
 import BasicAvatar from '@/components/shared/avatar';
-import useDetails from '@/hooks/pages/useDetails';
 import BasicChip from '@/components/shared/chip';
 import { PROFILE_URL, PROJECT_MODIFYL_URL } from '@/consts/routes';
 import Comments from '@/components/details/comments';
+import useProjectDetails from '@/pages/projects/useProjectDetails';
 
 export default function ProjectDetailPage() {
   const {
@@ -19,7 +19,7 @@ export default function ProjectDetailPage() {
     createdAt,
     onClick,
     isAuthor,
-  } = useDetails();
+  } = useProjectDetails();
 
   return (
     <Stack spacing={3}>
